+++ pytest_mqtt/mosquitto.py
@@ -80,6 +80,19 @@ def is_mosquitto_running(host: str, port: int) -> bool:
     return probe_tcp_connect(host, port)
 
 
+def pytest_addoption(parser) -> None:
+    parser.addoption("--mqtt-host", action="store", type=str, default="localhost", help="MQTT host name")
+    parser.addoption("--mqtt-port", action="store", type=int, default=1883, help="MQTT port number")
+
+
+@pytest.fixture(scope="session")
+def mqtt_settings(pytestconfig) -> MqttSettings:
+    return MqttSettings(
+        host=pytestconfig.getoption("--mqtt-host"),
+        port=pytestconfig.getoption("--mqtt-port"),
+    )
+
+
 @pytest.fixture(scope="session")
 def mosquitto(mqtt_settings: MqttSettings):
